# api/openapi.yaml
openapi: 3.1.0
info:
  title: Tindarr API
  version: 1.0.0-draft
  summary: Self-hosted server behind the Tindarr mobile app and web console.
  description: |
    Contract between a Tindarr server and its two clients: the mobile app and the
    web console the server serves under `/`. Written before the implementation; CI
    validates the server's responses against this file and rejects routes it does
    not document.

    Rules (see docs/adr/0006):
    - Additive changes only within v1. Clients ignore unknown fields and treat
      unknown enum values as "other".
    - Errors are `application/problem+json` with a stable `code`.
    - Slow work answers `202` with `retry_after_ms`; the client polls.
    - Every `429` carries `retry_after_ms` and a `Retry-After` header.
    - CORS is disabled: the console is same-origin.

    Every request (docs/auth.md, sections 1 and 2):
    - The `Host` header must be an IP literal, `localhost`, a name in
      `TINDARR_ALLOWED_HOSTS` or the host of `public_url`; anything else gets
      `400` `host_not_allowed` before routing, on every path but `/healthz`.
    - The client IP and scheme come from `X-Forwarded-For` and
      `X-Forwarded-Proto` only when the peer is in `TINDARR_TRUSTED_PROXIES`.
      The server's origin is `<scheme>://<Host>`.
    - Codes, handles and tokens travel in bodies, cookies or headers, never in a
      path or query string.

    Authentication (docs/adr/0009, docs/auth.md section 2), in this order:
...

...
  license:
...
servers:
...
security:
  - bearer: []
  - cookieSession: []
    csrfToken: []
...
paths:
...
  /api/v1/server/info:
    get:
      tags: [server]
      summary: Server identity, version and capabilities
      description: |
        Public. Returns no user data and no internal URL. `auth_methods` depends on
        the caller's client IP (`password_sign_in = lan_only`), so the response is
        never cached. Never calls the media server.
      operationId: getServerInfo
      security: []
      parameters:
        - name: Tindarr-Verify-Nonce
          in: header
          required: false
          description: |
            Sent only by this server to itself while it verifies a new `public_url`
            (docs/auth.md, section 10). When the nonce is one the server is waiting
            for, the response carries `public_url_proof`; otherwise the header is
            ignored.
          schema: { type: string, maxLength: 64 }
      responses:
        "200":
          description: Server information.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServerInfo" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/RateLimited" }

  # ---------------------------------------------------------------- setup
...
  /api/v1/auth/pair/preview:
    post:
      tags: [auth, pairing]
      summary: Show who a pairing code signs in, without using it
      description: |
        The app calls this after scanning a QR code, shows the host of the QR
        link prominently (punycode for internationalised names) with the names
        returned here, and asks the user to confirm before calling
        `POST /auth/pair`. Returns only what the app must display, and only for a
        `pending`, unexpired code.
      operationId: previewPairing
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { $ref: "#/components/schemas/PairingCode" }
      responses:
        "200":
          description: The pairing is valid.
          content:
            application/json:
              schema:
                type: object
                required: [server_name, user_name, expires_at]
                properties:
                  server_name: { type: string }
                  user_name: { type: string }
                  expires_at: { type: string, format: date-time }
        "400": { $ref: "#/components/responses/BadRequest" }
        "410":
          description: "`pairing_expired` (also for an unknown, used, requested or revoked code)."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/auth/pair:
    post:
      tags: [auth, pairing]
      summary: Ask to pair the app with a code from the console
      description: |
        Records the device and the PKCE `code_challenge`, and moves the pairing to
        `awaiting_approval`; the code then stops working for preview and pair. The
        app shows `confirmation_code`, which the console shows too, and polls
        `POST /auth/pair/complete` until the user approves or rejects the request in
        the console (docs/auth.md, section 9). Preview, pair and complete share a
        limit of 10 calls per minute per client IP, with a global slowdown.
      operationId: pairDevice
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, device, code_challenge]
              properties:
                code: { $ref: "#/components/schemas/PairingCode" }
                device: { $ref: "#/components/schemas/DeviceInput" }
                code_challenge: { $ref: "#/components/schemas/CodeChallenge" }
      responses:
        "202":
          description: Waiting for approval in the console.
          headers:
            Retry-After:
              description: Seconds, rounded up.
              schema: { type: integer }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PairingRequested" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "410":
          description: "`pairing_expired` (also for an unknown, used, requested or revoked code)."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/auth/pair/complete:
    post:
      tags: [auth, pairing]
      summary: Get the tokens once the console approved the pairing
      description: |
        Polled by the app every 2 seconds after `POST /auth/pair`. Once the pairing
        is approved, opens a `mobile` session for the user who created it and marks
        the pairing `completed`. Applies the checks of any authenticated request
        (user enabled, remote-access rule). Does not contact the media server, so
        the administrator flag is not re-synced (docs/adr/0010).
      operationId: completePairing
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, code_verifier]
              properties:
                code: { $ref: "#/components/schemas/PairingCode" }
                code_verifier: { $ref: "#/components/schemas/CodeVerifier" }
      responses:
        "200":
          description: Signed in.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResult" }
        "202": { $ref: "#/components/responses/Pending" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403":
          description: "`pairing_rejected` (rejected in the console), `account_disabled` or `remote_access_denied`."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "410":
          description: "`pairing_expired` (also for an unknown, completed or revoked code, a code never requested, or a wrong `code_verifier`)."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /api/v1/pairings:
    post:
      tags: [pairing, console]
      summary: Create a code to connect a phone (any signed-in user)
      description: |
        The code is bound to the caller, single-use, valid 5 minutes and stored
        hashed; it is only returned here. The console renders `link` as a QR code
        (SVG or canvas) with the host of `public_url` written next to it. At most 3
        pending pairings per user (`rate_limited` beyond).
      operationId: createPairing
      security:
        - cookieSession: []
          csrfToken: []
      responses:
        "201":
          description: Pairing created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NewPairing" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: "`public_url_not_set`: a media server administrator must set `public_url` first."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/pairings/{pairing_id}:
    get:
      tags: [pairing, console]
      summary: Status of one of the caller's pairings
      description: |
        Polled by the console. When `awaiting_approval`, shows the requesting
        device, its IP and the confirmation code the phone displays, so the user
        can approve or reject it; once `completed`, "Pixel 9 connected".
      operationId: getPairing
      security:
        - cookieSession: []
      parameters:
        - $ref: "#/components/parameters/PairingId"
      responses:
        "200":
          description: Pairing.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Pairing" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [pairing, console]
      summary: Cancel or reject a pairing, or revoke the session it opened
      description: |
        `pending` or `approved`: cancelled. `awaiting_approval`: rejected, and the
        app's next `POST /auth/pair/complete` gets `403` `pairing_rejected`.
        `completed`: the `mobile` session it opened is revoked.
      operationId: revokePairing
      security:
        - cookieSession: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/PairingId"
      responses:
        "204": { description: "Cancelled, rejected or revoked." }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/pairings/{pairing_id}/approve:
    post:
      tags: [pairing, console]
      summary: Approve the phone waiting on this pairing
      description: |
        Only for a pairing in `awaiting_approval`. The console asks the user to
        check that the phone shows the same `confirmation_code` first. The app's
        next `POST /auth/pair/complete` then gets the tokens.
      operationId: approvePairing
      security:
        - cookieSession: []
          csrfToken: []
      parameters:
        - $ref: "#/components/parameters/PairingId"
      responses:
        "200":
          description: Approved.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Pairing" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "`pairing_not_awaiting_approval` (no phone has asked yet, or already approved, completed or revoked)."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "410":
          description: "`pairing_expired`."
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }

  # ---------------------------------------------------------------- swipe
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from an app sign-in, pairing or refresh, valid 15 minutes. Mobile app only.
    cookieSession:
      type: apiKey
      in: cookie
      name: __Host-tindarr_session
      description: |
        Web console session (`kind` = `web`): opaque 256-bit token, stored hashed,
        `HttpOnly`, `Secure`, `SameSite=Strict`, `Path=/`, no `Domain`. Expires
        after 24 h idle or 7 days after sign-in. Named `tindarr_session`, without
        `Secure`, only on requests from a private client IP when
        `TINDARR_ALLOW_HTTP_CONSOLE=true` (docs/auth.md, section 2).
    setupSession:
      type: apiKey
      in: cookie
      name: __Host-tindarr_setup
      description: |
        Setup session from `POST /setup/claim` (`kind` = `setup`, no user,
        30 minutes absolute, same attributes as the web session cookie; named
        `tindarr_setup` under the plain-HTTP opt-in). Only one is valid at a time.
        Accepted by the setup operations, `GET /auth/web/session`, the web sign-ins
        while setup is pending, and Plex PINs of purpose `owner_token`. Never
        accepted where a web session is expected, and the reverse.
    preAuth:
      type: apiKey
      in: cookie
      name: __Host-tindarr_preauth
      description: |
        Not a session: a random value that binds the console's Plex PIN and Quick
        Connect sign-in handles to the browser that started them (docs/auth.md,
        section 5). Set by `POST /auth/plex/pins` and `POST /auth/quick-connect`
        when called by the console without `code_challenge`, if absent; 15 minutes;
        cleared by a successful web sign-in. Named `tindarr_preauth` under the
        plain-HTTP opt-in.
    csrfToken:
      type: apiKey
      in: header
...
  parameters:
...
    PairingId:
      name: pairing_id
      in: path
      required: true
      schema: { type: string }
...
  headers:
...
    RetryAfter:
      description: Seconds, rounded up.
      schema: { type: integer, minimum: 0 }
  responses:
...
    BadRequest:
      description: |
        `validation_error` (with `errors`; input values are never echoed),
        `bad_request` (malformed body), or `host_not_allowed` (the `Host` header is
        not an allowed host; any path but `/healthz`).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Unauthorized:
      description: |
        `unauthorized`: no credential this operation accepts (an `Authorization`
        header on an operation without `bearer`, or a cookie of another kind,
        counts as none), a revoked, idle-expired or past-lifetime session, a
        disabled or unlinked user. `token_expired`: the bearer access token is past
        `exp`; refresh (single-flight) and retry.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Forbidden:
      description: |
        `csrf_failed` (cookie request with an unsafe method: wrong or missing
        `X-CSRF-Token` or `Origin`) or `remote_access_denied` (the user's media
        server policy forbids remote access and the client IP is not private).
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    NotFound:
      description: "`not_found`, also for another user's resource."
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    RateLimited:
      description: |
        `rate_limited`, with `retry_after_ms`. Limits are per client IP (or per
        username, session or user where docs/auth.md section 8 says so); global
        thresholds only slow requests down, except the cap on outstanding sign-in
        handles.
      headers:
        Retry-After: { $ref: "#/components/headers/RetryAfter" }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Pending:
      description: Work in progress; ask again after `retry_after_ms`.
      headers:
        Retry-After: { $ref: "#/components/headers/RetryAfter" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Pending" }
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details with a stable machine-readable `code`.
      required: [type, title, status, code]
      properties:
        type: { type: string, format: uri-reference, default: about:blank }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code:
          type: string
          description: |
            Stable identifier. Known values:
            - request and credentials: validation_error, bad_request,
              host_not_allowed, unauthorized, token_expired, refresh_token_reused,
              csrf_failed, https_required, forbidden, admin_required,
              media_server_admin_required, reauth_required, remote_access_denied,
              not_found, method_not_allowed, rate_limited, internal_error;
            - setup: setup_required, setup_completed, invalid_setup_code,
              setup_session_required, setting_locked;
            - sign-in: invalid_credentials, account_disabled, not_a_server_user,
              password_sign_in_disabled, sign_in_method_unavailable, pin_expired,
              plex_pin_pending, plex_owner_required, plex_tv_unreachable,
              quick_connect_unavailable, quick_connect_expired,
              media_server_unreachable, media_server_changed;
            - pairing: pairing_expired, pairing_rejected,
              pairing_not_awaiting_approval, public_url_not_set,
              public_url_unverified;
            - connectors: connector_unreachable, connector_unauthorized,
              connector_unexpected_response, media_server_unsupported,
              media_server_required, secret_required;
            - swipe and admin: last_admin, streaming_region_not_set,
              metadata_unreachable, llm_not_configured, tmdb_not_configured,
              llm_auth_failed, llm_quota, llm_model_not_found, llm_unreachable,
              llm_invalid_output, daily_limit_reached, requests_not_configured,
              no_backend_user, title_not_offered, request_not_allowed,
              quota_exceeded, request_backend_error.
          examples: [llm_quota]
        retry_after_ms:
          type: integer
          minimum: 0
          description: Set on `rate_limited` and `daily_limit_reached`; the `Retry-After` header carries the same wait in seconds, rounded up.
        reason:
          type: string
          description: |
            Coarse cause for `public_url_unverified`, never a response body:
            `unreachable`, `tls_error`, `redirected`, `unexpected_response` or
            `proof_mismatch`.
          examples: [proof_mismatch]
        errors:
          type: array
          description: Field errors for `validation_error`.
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
    Pending:
      type: object
      required: [pending, retry_after_ms]
      properties:
        pending: { type: boolean, const: true }
        retry_after_ms: { type: integer, minimum: 250 }
    ServerInfo:
      type: object
      required: [name, version, api_version, min_app_version, setup_required, auth_methods, capabilities, tmdb_image_base_url]
      properties:
        name: { type: string, description: Server display name chosen by the admin. }
        version: { type: string, examples: ["1.0.0"] }
        api_version: { type: integer, const: 1 }
        min_app_version: { type: string, examples: ["1.0.0"] }
        setup_required: { type: boolean }
        media_server:
          oneOf:
            - type: "null"
            - type: object
              required: [kind]
              properties:
                kind: { $ref: "#/components/schemas/MediaServerKind" }
                name:
                  type: [string, "null"]
                  description: What the media server calls itself, read at the last successful connection test; null while it is unknown.
        auth_methods:
          type: array
          description: |
            Ways to sign in for this caller (docs/auth.md, section 11). Empty while
            `setup_required` is true, whatever is configured (the setup wizard
            reads `GET /setup/state` instead). Otherwise:
            - `password`: Jellyfin or Emby, unless `password_sign_in` is `disabled`,
              or `lan_only` and the caller's client IP is not private;
            - `quick_connect`: Jellyfin, when the cached `GET /QuickConnect/Enabled`
              says so (left out while unknown);
            - `plex_pin`: Plex;
            - `pairing`: once `public_url` is set.
          uniqueItems: true
          items: { $ref: "#/components/schemas/AuthMethod" }
        capabilities:
          type: array
          description: Feature flags the clients enable features on, e.g. `trailers`, `ratings`, `streaming_providers`, `requests`, `notifications`.
          items: { type: string }
        tmdb_image_base_url:
          type: string
          format: uri
          description: Base the app prepends to `*_path` image fields (size segment appended by the app).
          examples: ["https://image.tmdb.org/t/p/"]
        public_url_proof:
          type: string
          description: |
            base64url(HMAC-SHA256(key, nonce)) with the `tindarr/v1/public-url-proof`
            sub-key. Present only when the request carried a
            `Tindarr-Verify-Nonce` this server is currently waiting for.
      if:
        required: [setup_required]
        properties:
          setup_required: { const: true }
      then:
        properties:
          auth_methods: { maxItems: 0 }
    AuthMethod:
      type: string
      enum: [password, plex_pin, quick_connect, pairing]
...
    CodeVerifier:
      type: string
      description: PKCE code verifier (RFC 7636), kept by the app.
      pattern: "^[A-Za-z0-9._~-]{43,128}$"
    CodeChallenge:
      type: string
      description: "PKCE S256 challenge: base64url(SHA-256(`code_verifier`)), without padding."
      pattern: "^[A-Za-z0-9_-]{43}$"
...
    PairingRequested:
      type: object
      description: The pairing is `awaiting_approval`; poll `POST /auth/pair/complete`.
      required: [pending, retry_after_ms, confirmation_code, expires_at]
      properties:
        pending: { type: boolean, const: true }
        retry_after_ms: { type: integer, minimum: 250, examples: [2000] }
        confirmation_code: { $ref: "#/components/schemas/ConfirmationCode" }
        expires_at:
          type: string
          format: date-time
          description: The pairing's expiry, 5 minutes after it was created.
    ConfirmationCode:
      type: string
      description: |
        4 digits derived from the `code_challenge` (first 32 bits of its SHA-256,
        modulo 10 000, zero-padded). The app and the console show it; the user
        approves only if both match.
      pattern: "^[0-9]{4}$"
      examples: ["4821"]
    DeviceInput:
      type: object
      description: Sent by the app. Web sessions describe the browser from its User-Agent instead.
      required: [name, platform, app_version]
      properties:
        name: { type: string, maxLength: 80, examples: ["Pixel 9"] }
        platform: { type: string, enum: [android, ios] }
        app_version: { type: string, maxLength: 32 }
...
    PairingCode:
      type: string
      description: At least 128 bits, base64url.
      pattern: "^[A-Za-z0-9_-]{22,64}$"
...
    NewPairing:
      allOf:
        - $ref: "#/components/schemas/Pairing"
        - type: object
          required: [code, link]
          properties:
            code: { $ref: "#/components/schemas/PairingCode" }
            link:
              type: string
              description: "`tindarr://pair?server=<public_url>&code=<code>`, rendered as a QR code (SVG or canvas), with the host of `public_url` written next to it."
              examples: ["tindarr://pair?server=https%3A%2F%2Ftindarr.example.com&code=q7Xc0vW2dYk9LmN4pRs6Tu"]
    TokenPair:
      type: object
      required: [access_token, access_expires_at, refresh_token, refresh_expires_at]
      properties:
        access_token: { type: string }
        access_expires_at: { type: string, format: date-time }
        refresh_token: { type: string }
        refresh_expires_at: { type: string, format: date-time }
    AuthResult:
      allOf:
        - $ref: "#/components/schemas/TokenPair"
        - type: object
          required: [user, setup_completed_now]
          properties:
            user: { $ref: "#/components/schemas/User" }
            setup_completed_now:
              type: boolean
              description: True when this sign-in completed the first-run setup.
...
    MediaServerKind:
      type: string
      enum: [jellyfin, emby, plex]
...